WIP: Introducing cornerstoneTools v3 into Viewers
